+++ test/src/models/form_group_test.dart
@@ -167,8 +167,8 @@ void main() {
       expect(form.control('name').touched, false);
       expect(form.control('email').touched, false);
 
-      // When: touch the form
-      form.touch();
+      // When: marks all descendants to touched
+      form.markAllAsTouched();
 
       // Then: all controls are touched
       expect(form.control('name').touched, true);
@@ -187,7 +187,7 @@ void main() {
       expect(form.control('email').touched, true);
 
       // When: untouch the form
-      form.untouch();
+      form.markAsUntouched();
 
       // Then: all controls are untouched
       expect(form.control('name').touched, false);
@@ -261,7 +261,7 @@ void main() {
       });
 
       // When: disable group
-      form.disable();
+      form.markAsDisabled();
 
       // Then: children are disabled
       expect(form.control('name').disabled, true);
@@ -292,7 +292,7 @@ void main() {
       });
 
       // When: enable form
-      form.enable();
+      form.markAsEnabled();
 
       // Then: all controls are enabled
       expect(form.controls.values.every((control) => control.enabled), true);
@@ -318,7 +318,7 @@ void main() {
       });
 
       // When: disable invalid control
-      form.control('email').disable();
+      form.control('email').markAsDisabled();
 
       // Then: form is valid
       expect(form.valid, true);
@@ -333,7 +333,7 @@ void main() {
       });
 
       // When: enable invalid control
-      form.control('email').enable();
+      form.control('email').markAsEnabled();
 
       // Then: form is invalid
       expect(form.invalid, true, reason: 'form is valid');
@@ -427,8 +427,67 @@ void main() {
       form.resetState({});
 
       // Then: all controls has null value
-      expect(form.control('name').value, null);
-      expect(form.control('name').touched, false);
+      expect(form.control('name').value, null, reason: 'value is not null');
+      expect(form.control('name').touched, false, reason: 'control is touched');
+    });
+
+    test('Reset a group marks it as pristine', () {
+      // Given: a group
+      final form = FormGroup({
+        'name': FormControl(),
+      });
+
+      // When: marks it as dirty
+      form.markAsDirty();
+
+      // Expect: is dity
+      expect(form.dirty, true);
+
+      // When: resets the group
+      form.reset();
+
+      // Then: value of the control has the new initial value
+      expect(form.pristine, true);
+    });
+
+    test('Resets a group state marks it as pristine', () {
+      // Given: a group
+      final form = FormGroup({
+        'name': FormControl(),
+      });
+
+      // When: marks it as dirty
+      form.markAsDirty();
+
+      // Expect: form is dirty
+      expect(form.dirty, true);
+
+      // When: resets the group
+      form.resetState({});
+
+      // Then: all controls has null value
+      expect(form.dirty, false, reason: 'form is dirty');
+      expect(form.pristine, true, reason: 'form is dirty');
+    });
+
+    test('Resets a group state marks it as pristine', () {
+      // Given: a group
+      final form = FormGroup({
+        'name': FormControl(value: 'initial value'),
+      });
+
+      // When: marks it as dirty
+      form.markAsDirty();
+
+      // Expect: form is dirty
+      expect(form.dirty, true);
+
+      // When: resets the group
+      form.resetState({'name': null});
+
+      // Then: all controls has null value
+      expect(form.dirty, false, reason: 'form is pristine');
+      expect(form.pristine, true, reason: 'form is dirty');
     });
   });
 }