+++ test/src/widgets/reactive_radio_test.dart
@@ -151,7 +151,7 @@ void main() {
         await tester.pumpWidget(ReactiveRadioTestingWidget(form: form));
 
         // When: disable form
-        form.disable();
+        form.markAsDisabled();
         await tester.pump();
 
         // Then: the radio is disabled
@@ -178,7 +178,7 @@ void main() {
         await tester.pumpWidget(ReactiveRadioTestingWidget(form: form));
 
         // When: enable form
-        form.enable();
+        form.markAsEnabled();
         await tester.pump();
 
         // Then: the radio is enabled