+++ CHANGELOG.md
@@ -1,5 +1,15 @@
-# 4.1.0
+# 5.0.0
 
+## Breaking changes
+- Rename *touch()* and *untouch()* by *markAsTouched()* and *markAsUntouched()*
+- Rename *enable()* and *disable()* by *markAsEnabled()* and *markAsDisabled()*
+- Add **AbstractControl.markAllAsTouched** to mark all controls of a 
+**FromGroup** or a **FormArray** as touched. **AbstractControl.markAsTouched** doesn't 
+marks children as touched anymore.
+ 
+## Features
+
+- Add new control status *pristine* and *dirty*.
 - Add *ReactiveRadioListTile* widget.
 
 # 4.0.2